LayoutTests/imported/w3c/web-platform-tests/custom-elements/reactions/customized-builtins/HTMLOptGroupElement-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: testReflectBooleanAttributeWithDependentAttributes

Harness Error (FAIL), message = ReferenceError: Can't find variable: testReflectBooleanAttributeWithDependentAttributes
